uint data type support

+++ pim-func-sim/libpimsim/src/pimSim.cpp
@@ -752,6 +752,6 @@ template bool pimSim::pimBroadcast<int64_t>(PimObjId dest, int64_t value);
 template bool pimSim::pimRedSum<uint64_t>(PimObjId src, uint64_t* sum);
 template bool pimSim::pimRedSum<int64_t>(PimObjId src, int64_t* sum);
 
-template bool pimSim::pimRedSumRanged<uint64_t>(PimObjId src, unsigned idxBegin, unsigned idxEnd, uint64_t* sum);
-template bool pimSim::pimRedSumRanged<int64_t>(PimObjId src, unsigned idxBegin, unsigned idxEnd, int64_t* sum);
+template bool pimSim::pimRedSumRanged<uint64_t>(PimObjId src, uint64_t idxBegin, uint64_t idxEnd, uint64_t* sum);
+template bool pimSim::pimRedSumRanged<int64_t>(PimObjId src, uint64_t idxBegin, uint64_t idxEnd, int64_t* sum);
